Refactor code to move onClick function to a utility
